// backend/src/prisma/queries/rapor/interfaces/rapor.ts
export type UpdateRaporStatic = {
    // // static value murid
    // nama: string;
    // nis: string;
    // tempatLahir: string;
    // tanggalLahir: Date;
    // jenisKelamin: string;
    // agama: string;
    // anakKe: number;

    // // static value orang tua
    // namaAyah: string;
    // namaIbu: string;
    // pekerjaanAyah: string;
    // pekerjaanIbu: string;
    // jalan: string;
    // kelurahan: string;
    // kecamatan: string;
    // kota: string;
    // provinsi: string;
    // foto: string;

    // static value guru
    namaGuru: string;
    nipGuru: string;
    namaKapsek: string;
    nipKapsek: string;

    // static value rombel
    namaRombel: string;
    kelompokUsia: string;

    // static value sekolah
    namaSekolah: string;
    npsnSekolah: string;
    alamatSekolah: string;
    kodePosSekolah: string;
    noTeleponSekolah: string;
    kelurahanSekolah: string;
    kecamatanSekolah: string;
    kotaSekolah: string;
    provinsiSekolah: string;
    namaDisdikSekolah: string;
    logoSekolah: string;

    // static value semester
    namaSemester: string;
    semesterTahun: string;
    tanggalBagiRapor: Date;
};